requirements.txt
# Orchestrator dependencies
fastapi==0.115.*
uvicorn[standard]==0.34.*
python-statemachine==2.5.*
click==8.1.*
httpx==0.28.*
pyyaml==6.0.*
jinja2==3.1.*
docker==7.1.*
